+++ ext/rucaptcha/extconf.rb
@@ -1,4 +1,2 @@
 require 'mkmf'
-
-# dir_config(extension_name)
 create_makefile('rucaptcha/rucaptcha')