+++ CHANGELOG.md
@@ -6,8 +6,8 @@
 WARNING!: This version have so many break changes!
 
 - Use C ext instead of ImageMagick, now it's no dependencies!
-- New captcha style, because C code is complex, just generate black chars.
-- Remove `len`, `style`, `font_size`, `cache_limit` config key, no support now.
+- New captcha style.
+- Remove `len`, `font_size`, `cache_limit` config key, no support now.
 
 1.2.0
 -----